AzureFunctionApp/requirements.txt
psycopg2-binary
pyjwt==2.6.0
environ==1.0
django-environ
django==4.1.5
djangorestframework
whitenoise
django-webpack-loader
django-filter
openai
pandas
numpy
requests
datetime
azure-cognitiveservices-vision-computervision
PyPDF2
Pillow
pdf2image
opencv-python-headless
geopy
django-cloudinary-storage
cloudinary
azure-functions
azure-storage-blob
weasyprint